requirements.txt
flask>=3.0
gunicorn>=21.2
twilio>=9.0
elevenlabs>=1.0
google-cloud-aiplatform>=1.38
google-cloud-firestore>=2.14
google-cloud-logging>=3.8
requests>=2.31
httpx>=0.26.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-gcp-trace>=1.6
pytest>=8.0
